requirements.txt
langchain
langchain-openai
python-dotenv
ipykernel
beautifulsoup4
langchain-community
faiss-cpu
streamlit